refactor(perl-regex): split srp-focused modules (#0000)
